+++ src/base/state/sub_state.jl
@@ -1,39 +1,51 @@
 """
-    AbstractSubProblemSolverState <: AbstractManoptSolverState
+    has_sub_problem(::Type{<:AbstractManoptSolverState})
 
-An abstract type for solvers that involve a subsolver.
+Return whether a solver state type stores a sub task as the pair `sub_problem` and `sub_state`,
+see [`get_sub_problem`](@ref) and [`get_sub_state`](@ref).
+This is `false` by default; a state that stores such a pair declares it as
+`has_sub_problem(::Type{<:MyState}) = true`.
 """
-abstract type AbstractSubProblemSolverState <: AbstractManoptSolverState end
+has_sub_problem(::Type{<:AbstractManoptSolverState}) = false
 
 @doc """
-    get_sub_problem(ams::AbstractSubProblemSolverState)
+    get_sub_problem(ams::AbstractManoptSolverState)
 
-Access the sub problem of a solver state that involves a sub optimization task.
-By default this returns `ams.sub_problem`.
+Access the sub problem of a solver state that involves a sub optimization task,
+see [`has_sub_problem`](@ref). By default this returns `ams.sub_problem`.
 """
-get_sub_problem(ams::AbstractSubProblemSolverState) = ams.sub_problem
+get_sub_problem(ams::AbstractManoptSolverState) = _get_sub_problem(ams, Val(has_sub_problem(typeof(ams))))
+_get_sub_problem(ams::AbstractManoptSolverState, ::Val{true}) = ams.sub_problem
+function _get_sub_problem(ams::AbstractManoptSolverState, ::Val{false})
+    return error("The state $(typeof(ams)) does not store a sub problem.")
+end
 
 @doc """
-    get_sub_state(ams::AbstractSubProblemSolverState)
+    get_sub_state(ams::AbstractManoptSolverState)
 
-Access the sub state of a solver state that involves a sub optimization task.
-By default this returns `ams.sub_state`.
+Access the sub state of a solver state that involves a sub optimization task,
+see [`has_sub_problem`](@ref). By default this returns `ams.sub_state`.
 """
-get_sub_state(ams::AbstractSubProblemSolverState) = ams.sub_state
-
-"""
-    set_parameter!(ams::AbstractSubProblemSolverState, :SubProblem, args...)
-
-Set certain values specified by `args...` to the sub problem.
-"""
-function set_parameter!(ams::AbstractSubProblemSolverState, ::Val{:SubProblem}, args...)
-    return set_parameter!(get_sub_problem(ams), args...)
+get_sub_state(ams::AbstractManoptSolverState) = _get_sub_state(ams, Val(has_sub_problem(typeof(ams))))
+_get_sub_state(ams::AbstractManoptSolverState, ::Val{true}) = ams.sub_state
+function _get_sub_state(ams::AbstractManoptSolverState, ::Val{false})
+    return error("The state $(typeof(ams)) does not store a sub state.")
 end
-"""
-    set_parameter!(ams::AbstractSubProblemSolverState, :SubState, args...)
 
-Set certain values specified by `args...` to the sub state.
 """
-function set_parameter!(ams::AbstractSubProblemSolverState, ::Val{:SubState}, args...)
-    return set_parameter!(get_sub_state(ams), args...)
+    _set_sub_parameter!(ams::AbstractManoptSolverState, ::Val{has}, ::Val{e}, args...)
+
+The part of the fallback `set_parameter!(ams, Val(e), args...)` that concerns a sub task:
+if the state stores one (`has` is `true`, see [`has_sub_problem`](@ref)), `e = :SubProblem`
+passes `args...` on to `set_parameter!` of the sub problem and `e = :SubState` to the one of
+the sub state; every other combination leaves the state unchanged.
+"""
+function _set_sub_parameter!(ams::AbstractManoptSolverState, ::Val{true}, ::Val{:SubProblem}, args...)
+    set_parameter!(get_sub_problem(ams), args...)
+    return ams
 end
+function _set_sub_parameter!(ams::AbstractManoptSolverState, ::Val{true}, ::Val{:SubState}, args...)
+    set_parameter!(get_sub_state(ams), args...)
+    return ams
+end
+_set_sub_parameter!(ams::AbstractManoptSolverState, ::Val, ::Val, args...) = ams